chore: move SwiGLU network to torsos file
