+++ mava/networks/mat_network.py
@@ -15,31 +15,18 @@
 from typing import Tuple
 
 import chex
-import jax
 import jax.numpy as jnp
 from flax import linen as nn
 from flax.linen.initializers import orthogonal
 
 from mava.networks.attention import SelfAttention
+from mava.networks.torsos import SwiGLU
 from mava.networks.utils.mat.decode import (
     continuous_autoregressive_act,
     continuous_parallel_act,
     discrete_autoregressive_act,
     discrete_parallel_act,
 )
-
-
-class SwiGLU(nn.Module):
-    ffn_dim: int
-    embed_dim: int
-
-    def setup(self) -> None:
-        self.W_1 = self.param("W_1", nn.initializers.zeros, (self.embed_dim, self.ffn_dim))
-        self.W_G = self.param("W_G", nn.initializers.zeros, (self.embed_dim, self.ffn_dim))
-        self.W_2 = self.param("W_2", nn.initializers.zeros, (self.ffn_dim, self.embed_dim))
-
-    def __call__(self, x: chex.Array) -> chex.Array:
-        return (jax.nn.swish(x @ self.W_G) * (x @ self.W_1)) @ self.W_2
 
 
 class EncodeBlock(nn.Module):
